% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{usoc_compile}
\alias{usoc_compile}
\title{Compile Understanding Society SPSS files to one dataset}
\usage{
usoc_compile(directory, extra_mappings = NULL, save_variables_report = TRUE)
}
\arguments{
\item{directory}{Directory with \verb{bhps/} and \verb{ukhls/}}

\item{extra_mappings}{function with extra columns to compile}

\item{save_variables_report}{Save a csv with a summary of variables?}
}
\description{
Compiles all files from BHPS and UKHLS into one file
}
\details{
To automatically save this file in a directory to be used
between projects, set a system variable called \code{DATA_DIRECTORY}
pointing at a folder.

To do that just for R sessions, add the line below to .Rprofile:

\code{Sys.setenv(DATA_DIRECTORY = "~/path/to/folder")}
}
