% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{usoc_convert}
\alias{usoc_convert}
\title{Convert USOC files to RDS}
\usage{
usoc_convert(usoc_directory, new_directory, filter_files = "indresp")
}
\arguments{
\item{usoc_directory}{Directory of top level USOC data (where "bhps" and "ukhls" are)}

\item{new_directory}{Directory where RDS files will be saved}

\item{filter_files}{regex to narrow files - by default}
}
\description{
Convert USOC files to RDS
}
