% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{usoc_load}
\alias{usoc_load}
\title{Load USOC from DATA_DIRECTORY folder}
\usage{
usoc_load(file = "indresp", as.data.table = TRUE, ...)
}
\arguments{
\item{file}{Which file to load? By default indresp.}

\item{as.data.table}{Passed to read_fst}

\item{...}{Passed to fst::read_fst}
}
\description{
To be able to use one file across many projects,
set an environment variable called \code{DATA_DIRECTORY} with a folder path,
and use this function to load the data in.
}
